requirements.txt
psycopg2-binary==2.9.9
pytest==7.4.3
python-dotenv==1.0.0
psutil==5.9.6
pyyaml==6.0.1
# Linting and type checking (optional - for development)
mypy==1.7.1
ruff==0.1.6
pylint==4.0.4
pyright==1.1.407
types-psycopg2==2.9.21.12


